agent-py/pyproject.toml
[project]
name = "agent-py"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk>=0.2.11",
    "fastapi>=0.116.0",
    "google-adk>=1.5.0",
]
